@@ -14,7 +14,7 @@
 	<name>magiadoslivros</name>
 	<description>Demo project for Spring Boot</description>
 	<properties>
-		<java.version>11</java.version>
+		<java.version>17</java.version>
 	</properties>
 	<dependencies>
 		<dependency>
@@ -74,7 +74,6 @@
     		<artifactId>springfox-swagger-ui</artifactId>
     		<version>3.0.0</version>
 		</dependency>
-
 
 	</dependencies>
 
